+++ R/crew_async.R
@@ -83,7 +83,7 @@ crew_class_async <- R6::R6Class(
       mirai::daemons(
         n = private$.workers,
         dispatcher = FALSE,
-        autoexit = crew_signal_terminate(),
+        autoexit = crew_terminate_signal(),
         resilience = FALSE,
         idletime = 60000,
         .compute = private$.instance