+++ R/crew_launcher.R
@@ -444,7 +444,7 @@ crew_class_launcher <- R6::R6Class(
         (8L * as.integer(isTRUE(private$.garbage_collection)))
       list(
         url = socket,
-        autoexit = crew_signal_terminate(),
+        autoexit = crew_terminate_signal(),
         cleanup = cleanup,
         output = TRUE,
         maxtasks = private$.tasks_max,